backend/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# eQuant Backend — Python Dependencies
# Python 3.10+  |  pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────────────────

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1
gunicorn==22.0.0           # Production multi-worker launcher

# ── Data validation ───────────────────────────────────────────────────────────
pydantic==2.7.4
pydantic-settings==2.3.4   # Typed env-var config (BaseSettings)

# ── Async HTTP + scraping ─────────────────────────────────────────────────────
httpx==0.27.0              # Async HTTP client (data fetcher + Supabase REST calls)
beautifulsoup4==4.12.3     # HTML parsing for HiStock scraper
lxml==5.2.2                # Fast HTML/XML parser backend for BeautifulSoup

# ── Auth (JWT) ────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0  # JWT decode + JWKS key construction

# ── AI / LLM ─────────────────────────────────────────────────────────────────
openai==1.35.13            # OpenAI-compatible SDK (GPT-4o, DeepSeek, Qwen …)

# ── Database ─────────────────────────────────────────────────────────────────
supabase==2.5.3            # Supabase Python client (auth + database)

# ── Data processing (Quantitative core — Pandas-based MVP) ───────────────────
pandas==2.2.2
numpy==1.26.4

# ── Quantitative libraries (Phase 2 — uncomment when ready) ──────────────────
# NOTE: requires build-essential + cmake + libopenblas-dev (pre-installed by
# .devcontainer/setup.sh).  Keep commented during early development to
# speed up container build time.
#
# pyqlib==0.9.6            # Qlib data processing + feature engineering
# finrl==0.3.7             # FinRL environment module for strategy backtesting

# ── Dev / testing ─────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
httpx                      # Also used as async test client (TestClient via HTTPX)
